+++ cube/swiss/source/gui/IPLFontWrite.h
@@ -26,6 +26,7 @@ extern char txtbuffer[2048];
 void init_font(void);
 void drawString(int x, int y, const char *string, float scale, int align, GXColor fontColor);
 void drawStringMedium(int x, int y, const char *string, float scale, int align, GXColor fontColor);
+void drawStringMediumUntinted(int x, int y, const char *string, float scale, int align, GXColor fontColor);
 void drawStringWithCaret(int x, int y, const char *string, float scale, int align, GXColor fontColor, int caretPosition, GXColor caretColor);
 void drawStringEllipsis(int x, int y, const char *string, float scale, int align, GXColor fontColor, bool rotateVertical, int maxSize);
 int GetFontHeight(float scale);